requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.9
Django==5.1.4
django-jazzmin==3.0.1
fastapi==0.115.6
filelock==3.16.1
h11==0.14.0
idna==3.10
kombu==5.4.2
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
